% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{langApp}
\alias{langApp}
\title{langApp}
\usage{
langApp(...)
}
\arguments{
\item{...}{this param is for other unspecified params.}
}
\description{
Loads a langApp
}
\author{
Krzysztof Solowiej \href{mailto:krsolowiej@gmail.com}{krsolowiej@gmail.com}
}
